refactored fixed new *.page.tsx pages file-routes
